make Tensor backend agnostic (#303)
* sketch tensor refactor

* port all torch.Tensor to tensor.Tensor

* switch adjoint and gaussian to tensor.Tensor

* add test_tensor, pass test_adjoint

* add ops.finfo

* fix some gaussian tests

* most tests are passed now

* add backend to most tensor tests

* add backend until einsum tests

* add backend to all tests

* fix other tests

* force Tensor convert np.generic to np.ndarray

* temp save to debug

* fix final bug

* update docs of tensor

* update examples

* add comments:
funsor.torch.function -> funsor.function; Tensor.new_arange + Tensor.materialize

* remove arange from funsor.__init__

* nit

* remove debugging code

* oppss, lint
