+++ funsor/__init__.py
@@ -6,7 +6,7 @@ from funsor.integrate import Integrate
 from funsor.interpreter import reinterpret
 from funsor.sum_product import MarkovProduct
 from funsor.terms import Cat, Funsor, Independent, Lambda, Number, Slice, Stack, Variable, of_shape, to_data, to_funsor
-from funsor.torch import Tensor, arange
+from funsor.tensor import Tensor, function
 from funsor.util import pretty, quote
 
 from . import (
@@ -47,7 +47,6 @@ __all__ = [
     'Variable',
     'adjoint',
     'affine',
-    'arange',
     'backward',
     'bint',
     'cnf',
@@ -56,6 +55,7 @@ __all__ = [
     'domains',
     'einsum',
     'find_domain',
+    'function',
     'gaussian',
     'integrate',
     'interpreter',
