+++ test/test_delta.py
@@ -7,9 +7,9 @@ import torch
 import funsor.ops as ops
 from funsor.delta import Delta
 from funsor.domains import reals
+from funsor.tensor import Tensor
 from funsor.terms import Number, Variable
 from funsor.testing import assert_close, check_funsor
-from funsor.torch import Tensor
 
 
 def test_eager_subs_variable():
