+++ test/test_joint.py
@@ -15,9 +15,9 @@ from funsor.gaussian import Gaussian
 from funsor.integrate import Integrate
 from funsor.interpreter import interpretation
 from funsor.montecarlo import monte_carlo_interpretation
+from funsor.tensor import Tensor
 from funsor.terms import Number, Variable, eager, moment_matching
 from funsor.testing import assert_close, random_gaussian, random_tensor, xfail_if_not_implemented
-from funsor.torch import Tensor
 
 
 def id_from_inputs(inputs):
